Переезд сотрудников в отдельный контроллер
